Migrate to spray-json
lift-json heavily relies on some reflection stuffs, it also has a lot of
dependencies and slow release cycles.
This dependency for something like a json library turned out to be
really hard to manage as diffson, and consequently sohva, and the
\BlueLaTeX cannot migrate to scala 2.11 which offers nice new stuffs.

By migrating to a library like spray-json, which follows more tightly
scala new releases and does not use reflection at all, we can be more
reactive to publish new versions of diffson, and we can think about
porting diffson to Scala.js more easily.